app/pyproject.toml
[project]
name = "recommender-app"
version = "1.0.0"
description = "Recommender Accelerator demo app"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.110.0",
    "uvicorn>=0.27.0",
    "psycopg2-binary>=2.9.0",
    "pydantic>=2.0.0",
]
